+++ cmd/pkg/install.go
@@ -3,7 +3,6 @@ package pkg
 import (
 	"fmt"
 	"github.com/oslokommune/ok/pkg/pkg/install"
-	"github.com/oslokommune/ok/pkg/pkg/install/select_pkg"
 	"github.com/spf13/cobra"
 )
 
@@ -31,7 +30,7 @@ BASE_URL=../boilerplate/terraform ok install networking my-app
 	SilenceErrors: true,
 	RunE: func(cmd *cobra.Command, stacks []string) error {
 		if installSelectFlag {
-			answer, err := select_pkg.Run(PackagesManifestFilename)
+			answer, err := _select.Run(PackagesManifestFilename)
 			if err != nil {
 				return fmt.Errorf("selecting package: %w", err)
 			}
