+++ pkg/pkg/install/interactive/item_delegate.go
@@ -1,4 +1,4 @@
-package select_pkg
+package interactive
 
 import (
 	"fmt"
