+++ pkg/pkg/install/interactive/model.go
@@ -1,4 +1,4 @@
-package select_pkg
+package interactive
 
 import (
 	"github.com/charmbracelet/bubbles/list"
